Drop retry logic from autoscaler
This commit removes the Knative components such as rewinder and the
replayRoundTripper. This was done to move the retry logic into the rio
controller and the gateway controller. This allows for using envoy to
program the retry logic via the Glood virtualService.

This was done to increase the observability of the system and open the
option of cutomizable timeouts at a later point in time.